+++ internal/cmd/drive_upload.go
@@ -1,6 +1,7 @@
 package cmd
 
 import (
+	"bytes"
 	"context"
 	"fmt"
 	"io"
@@ -34,11 +35,11 @@ func (c *DriveUploadCmd) Run(ctx context.Context, flags *RootFlags) error {
 		return err
 	}
 
-	file, err := os.Open(opts.localPath)
+	media, err := openDriveUploadMedia(opts, c.KeepFrontmatter)
 	if err != nil {
 		return err
 	}
-	defer file.Close()
+	defer media.Close()
 
 	_, svc, err := requireDriveService(ctx, flags)
 	if err != nil {
@@ -46,9 +47,9 @@ func (c *DriveUploadCmd) Run(ctx context.Context, flags *RootFlags) error {
 	}
 
 	if opts.replaceFileID == "" {
-		return runDriveCreateUpload(ctx, svc, file, opts)
+		return runDriveCreateUpload(ctx, svc, media, opts)
 	}
-	return runDriveReplaceUpload(ctx, svc, file, opts)
+	return runDriveReplaceUpload(ctx, svc, media, opts)
 }
 
 func prepareDriveUpload(c *DriveUploadCmd) (driveUploadOptions, error) {
@@ -92,6 +93,30 @@ func prepareDriveUpload(c *DriveUploadCmd) (driveUploadOptions, error) {
 	}
 
 	return opts, nil
+}
+
+func driveUploadShouldStripMarkdownFrontmatter(opts driveUploadOptions, keepFrontmatter bool) bool {
+	return !keepFrontmatter && opts.convert && opts.mimeType == mimeTextMarkdown
+}
+
+func openDriveUploadMedia(opts driveUploadOptions, keepFrontmatter bool) (io.ReadCloser, error) {
+	file, err := os.Open(opts.localPath)
+	if err != nil {
+		return nil, err
+	}
+	if !driveUploadShouldStripMarkdownFrontmatter(opts, keepFrontmatter) {
+		return file, nil
+	}
+
+	data, readErr := io.ReadAll(file)
+	closeErr := file.Close()
+	if readErr != nil {
+		return nil, readErr
+	}
+	if closeErr != nil {
+		return nil, closeErr
+	}
+	return io.NopCloser(bytes.NewReader(stripYAMLFrontmatter(data))), nil
 }
 
 func runDriveCreateUpload(ctx context.Context, svc *drive.Service, file io.Reader, opts driveUploadOptions) error {